+++ OptimizelySDK/Config/HttpProjectConfigManager.cs
@@ -1,5 +1,5 @@
 ﻿/* 
- * Copyright 2019, Optimizely
+ * Copyright 2019-2020, Optimizely
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,11 +28,17 @@ namespace OptimizelySDK.Config
     {
         private string Url;
         private string LastModifiedSince = string.Empty;
-
+        private string DatafileAccessToken = string.Empty;
         private HttpProjectConfigManager(TimeSpan period, string url, TimeSpan blockingTimeout, bool autoUpdate, ILogger logger, IErrorHandler errorHandler) 
             : base(period, blockingTimeout, autoUpdate, logger, errorHandler)
         {
             Url = url;
+        }
+
+        private HttpProjectConfigManager(TimeSpan period, string url, TimeSpan blockingTimeout, bool autoUpdate, ILogger logger, IErrorHandler errorHandler, string datafileAccessToken)
+            : this(period, url, blockingTimeout, autoUpdate, logger, errorHandler)
+        {
+            DatafileAccessToken = datafileAccessToken;
         }
 
         public Task OnReady()
@@ -41,20 +47,44 @@ namespace OptimizelySDK.Config
         }
 
 #if !NET40 && !NET35
-        private static System.Net.Http.HttpClient Client;
+        // HttpClient wrapper class which can be used to mock HttpClient for unit testing.
+        public class HttpClient
+        {
+            private System.Net.Http.HttpClient Client;
+
+            public HttpClient()
+            {
+                Client = new System.Net.Http.HttpClient(GetHttpClientHandler());
+            }
+
+            public HttpClient(System.Net.Http.HttpClient httpClient) : this()
+            {
+                if (httpClient != null) {
+                    Client = httpClient;
+                }
+            }
+
+            public static System.Net.Http.HttpClientHandler GetHttpClientHandler()
+            {
+                var handler = new System.Net.Http.HttpClientHandler() {
+                    AutomaticDecompression = System.Net.DecompressionMethods.GZip | System.Net.DecompressionMethods.Deflate
+                };
+
+                return handler;
+            }
+
+            public virtual Task<System.Net.Http.HttpResponseMessage> SendAsync(System.Net.Http.HttpRequestMessage httpRequestMessage)
+            {
+                return Client.SendAsync(httpRequestMessage);
+            }
+        }
+
+        private static HttpClient Client;
+
         static HttpProjectConfigManager()
         {
-            Client = new System.Net.Http.HttpClient(GetHttpClientHandler());
-        }
-
-        public static System.Net.Http.HttpClientHandler GetHttpClientHandler()
-        {
-            var handler = new System.Net.Http.HttpClientHandler() {
-                AutomaticDecompression = System.Net.DecompressionMethods.GZip | System.Net.DecompressionMethods.Deflate
-            };
-
-            return handler;
-        }
+            Client = new HttpClient();
+        }        
 
         private string GetRemoteDatafileResponse()
         {
@@ -66,6 +96,10 @@ namespace OptimizelySDK.Config
             // Send If-Modified-Since header if Last-Modified-Since header contains any value.
             if (!string.IsNullOrEmpty(LastModifiedSince))
                 request.Headers.Add("If-Modified-Since", LastModifiedSince);
+
+            if (!string.IsNullOrEmpty(DatafileAccessToken)) {
+                request.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", DatafileAccessToken);
+            }
 
             var httpResponse =  Client.SendAsync(request);
             httpResponse.Wait();
@@ -136,11 +170,12 @@ namespace OptimizelySDK.Config
             private readonly TimeSpan DEFAULT_PERIOD = TimeSpan.FromMinutes(5);
             private readonly TimeSpan DEFAULT_BLOCKINGOUT_PERIOD = TimeSpan.FromSeconds(15);
             private readonly string DEFAULT_FORMAT = "https://cdn.optimizely.com/datafiles/{0}.json";
-
+            private readonly string DEFAULT_AUTHENTICATED_DATAFILE_FORMAT = "https://config.optimizely.com/datafiles/auth/{0}.json";
             private string Datafile;
+            private string DatafileAccessToken;            
             private string SdkKey;
             private string Url;
-            private string Format;
+            private string Format;            
             private ILogger Logger;
             private IErrorHandler ErrorHandler;
             private TimeSpan Period;
@@ -174,6 +209,14 @@ namespace OptimizelySDK.Config
 
                 return this;
             }
+#if !NET40 && !NET35
+            public Builder WithAccessToken(string accessToken)
+            {
+                this.DatafileAccessToken = accessToken;
+
+                return this;
+            }
+#endif
 
             public Builder WithUrl(string url)
             {
@@ -260,15 +303,18 @@ namespace OptimizelySDK.Config
                     ErrorHandler = new DefaultErrorHandler();
 
                 if (string.IsNullOrEmpty(Format)) {
-                    Format = DEFAULT_FORMAT;
+
+                    if (string.IsNullOrEmpty(DatafileAccessToken)) {
+                        Format = DEFAULT_FORMAT;
+                    } else {
+                        Format = DEFAULT_AUTHENTICATED_DATAFILE_FORMAT;
+                    }
                 }
 
-                if (string.IsNullOrEmpty(Url) && string.IsNullOrEmpty(SdkKey))
-                {
-                    ErrorHandler.HandleError(new Exception("SdkKey cannot be null"));                    
-                }
-                else if (!string.IsNullOrEmpty(SdkKey))
-                {
+                if (string.IsNullOrEmpty(Url)) {
+                    if (string.IsNullOrEmpty(SdkKey)) {
+                        ErrorHandler.HandleError(new Exception("SdkKey cannot be null"));
+                    }
                     Url = string.Format(Format, SdkKey);
                 }
 
@@ -290,7 +336,7 @@ namespace OptimizelySDK.Config
                 }
                     
 
-                configManager = new HttpProjectConfigManager(Period, Url, BlockingTimeoutSpan, AutoUpdate, Logger, ErrorHandler);
+                configManager = new HttpProjectConfigManager(Period, Url, BlockingTimeoutSpan, AutoUpdate, Logger, ErrorHandler, DatafileAccessToken);
 
                 if (Datafile != null)
                 {